Prevent issue which interfered with generation of attribs.
In some cases, the first item returned by contents() was a text node, when we really wanted the div next to the text node. We now use children() instead of contents() to ensure that we're getting elements, not text nodes.

One side effect is that if there are text nodes directly in .media-content, they will be excluded from the JSON. I'm not aware of any visualizations that actually have such markup, though.